put sync in scm folder
